Extracted page routes from API constants and replaced hardcoded routes with constants
