+++ lib/frontend/index.ts
@@ -3,4 +3,5 @@ export * from "./compareProps"
 export * from "./constants"
 export * from "./CSSProperties"
 export * from "./IconNames"
+export * from "./routes"
 export * from "./types"
