+++ pages/index.tsx
@@ -4,7 +4,8 @@ import React from "react"
 import { Button } from "../components/action"
 import { Card, CardAction, CardBody, CardTitle } from "../components/card"
 import { Overlay } from "../components/layout"
-import { ApiRoutes, fetchJson, PageRoutes } from "../lib/api"
+import { ApiRoutes, fetchJson } from "../lib/api"
+import { PageRoutes } from "../lib/frontend"
 import { useUser } from "../lib/hooks"
 
 const Landing: NextPage = () => {
